Revert "Remove unnecessary castings"
This reverts commit 022739b727fa504e52c415d4196de0bbb68fae8a.

+++ src/libai/nn/NeuralNetwork.java
@@ -167,7 +167,7 @@ public abstract class NeuralNetwork implements Serializable {
 				error += Math.pow(answers[i + offset].position(j, 0) - Y.position(j, 0), 2);
 		}
 
-		return error / length;
+		return error / (double) length;
 	}
 
 	/**
