+++ src/libai/nn/unsupervised/Hebb.java
@@ -183,7 +183,7 @@ public class Hebb extends NeuralNetwork {
 				error += Math.pow(Y.position(j, 0) - X.position(j, 0), 2);//0-1=1, 0-0=0, 1-1=0, 1-0=1
 		}
 
-		return error / (length * patterns[0].getRows());
+		return error / (double) (length * patterns[0].getRows());
 	}
 
 	/**
